arm: samples: tfm: Additional crypto examples
This commit adds a dedicated file for psa_crypto examples.
