+++ samples/tfm_integration/psa_level_1/src/psa_crypto.h
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) 2019 Linaro Limited
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <zephyr.h>
+
+#include "psa/error.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void crp_test_rng(void);
+void crp_test_sha256(void);
+
+#ifdef __cplusplus
+}
+#endif
